feat(settings): Add shipping rate method selection to settings page
